`nil' is now mapped to `()' at the lowest-level.
The symbol name `nil' has been added to ().

The generic function `equal' has been replaced by `binary=' to avoid frustrating
inconsistencies in the operation of {=, <, >, <=' >=}.  The only difference
between the definition of `equal' and `binary=' is the latter applies type
coercion whereas `equal' was defined not two.  If this property is an important
feature `equal' could be reinstated as a function wrapper around `binary=' which
does a type-consistency check.

`binary=' is now a "proper predicate" according to the definition and returns
true rather than one of the arguments.  This change has been propagated down to
the lowest level: `fpi-equal' within the corresponding byte-code.

The update to the definition document for these changes is not yet complete.